// src/app/interfaces/guia.ts
export interface Guia {
    id?: string;
    pacienteId?: string;
    nomePaciete?: string;
    nomePsicologo?: string;
    convenio?: string;
    qtdeSessoes?: string;
    dataInicio?: string;
    dataInicioStamp?: number;
    dataFim?: string;
    dataFimStamp?: number;
    diaSemana?: string;
    hora?: string;
    crmMedico?: string;
    numeroGuia?: string;
    userId?: string;
}